:ambulance: fix issues caused by ui rework merge
