+++ src/Components/MoreInfo.js
@@ -7,7 +7,7 @@ import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import config from './config.json';
+import config from '../config.json';
 
 const styles = theme => ({
 });
